requirements.txt
asgiref==3.2.7
Babel==2.8.0
dj-database-url==0.5.0
Django==3.0.4
django-filter==2.2.0
django-heroku==0.3.1
django-phonenumber-field==4.0.0
django-summernote==0.8.11.6
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
gunicorn==20.0.4
Markdown==3.2.1
phonenumbers==8.12.0
Pillow==7.0.0
psycopg2==2.8.4
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.1
whitenoise==5.0.1
